Free GPU mem, remove cargo feature, change type to support larger size evals

+++ icicle/include/fri/fri.cuh
@@ -32,7 +32,7 @@ namespace fri {
    * evaluations by half.
    */
   template <typename S, typename E>
-  cudaError_t fold_line(E* eval, S* domain_xs, E alpha, E* folded_eval, int n, FriConfig& cfg);
+  cudaError_t fold_line(E* eval, S* domain_xs, E alpha, E* folded_eval, uint64_t n, FriConfig& cfg);
 
   /**
    * @brief Folds a layer of FRI evaluations from a circle into a line.
@@ -57,7 +57,7 @@ namespace fri {
    */
 
   template <typename S, typename E>
-  cudaError_t fold_circle_into_line(E* eval, S* domain_ys, E alpha, E* folded_eval, int n, FriConfig& cfg);
+  cudaError_t fold_circle_into_line(E* eval, S* domain_ys, E alpha, E* folded_eval, uint64_t n, FriConfig& cfg);
 
 } // namespace fri
